[Behat] Move shared storage to behat services
