+++ src/Sylius/Behat/Context/Ui/Admin/ManagingOrdersContext.php
@@ -20,7 +20,7 @@ use Sylius\Behat\Service\SecurityServiceInterface;
 use Sylius\Component\Core\Model\CustomerInterface;
 use Sylius\Component\Core\Model\OrderInterface;
 use Sylius\Component\Core\Model\UserInterface;
-use Sylius\Component\Core\Test\Services\SharedStorageInterface;
+use Sylius\Behat\Service\SharedStorageInterface;
 use Webmozart\Assert\Assert;
 
 /**
