+++ src/Sylius/Behat/Resources/config/services/contexts/transform.xml
@@ -44,7 +44,7 @@
         <service id="sylius.behat.context.transform.customer" class="Sylius\Behat\Context\Transform\CustomerContext" scope="scenario">
             <argument type="service" id="sylius.repository.customer" container="symfony" />
             <argument type="service" id="sylius.factory.customer" container="symfony" />
-            <argument type="service" id="sylius.behat.shared_storage" container="symfony" />
+            <argument type="service" id="sylius.behat.shared_storage" />
             <tag name="sylius.behat.context" />
         </service>
 
@@ -104,7 +104,7 @@
         </service>
 
         <service id="sylius.behat.context.transform.shared_storage" class="Sylius\Behat\Context\Transform\SharedStorageContext" scope="scenario">
-            <argument type="service" id="sylius.behat.shared_storage" container="symfony" />
+            <argument type="service" id="sylius.behat.shared_storage" />
             <tag name="sylius.behat.context" />
         </service>
 
@@ -134,7 +134,7 @@
         </service>
 
         <service id="sylius.behat.context.transform.user" class="Sylius\Behat\Context\Transform\UserContext" scope="scenario">
-            <argument type="service" id="sylius.behat.shared_storage" container="symfony" />
+            <argument type="service" id="sylius.behat.shared_storage" />
             <tag name="sylius.behat.context" />
         </service>
