+++ src/Sylius/Behat/spec/Service/SharedStorageSpec.php
@@ -9,13 +9,13 @@
  * file that was distributed with this source code.
  */
 
-namespace spec\Sylius\Component\Core\Test\Services;
+namespace spec\Sylius\Behat\Service;
 
 use PhpSpec\ObjectBehavior;
 use Sylius\Component\Core\Model\ChannelInterface;
 use Sylius\Component\Core\Model\ProductInterface;
-use Sylius\Component\Core\Test\Services\SharedStorage;
-use Sylius\Component\Core\Test\Services\SharedStorageInterface;
+use Sylius\Behat\Service\SharedStorage;
+use Sylius\Behat\Service\SharedStorageInterface;
 
 /**
  * @mixin SharedStorage
@@ -26,7 +26,7 @@ class SharedStorageSpec extends ObjectBehavior
 {
     function it_is_initializable()
     {
-        $this->shouldHaveType('Sylius\Component\Core\Test\Services\SharedStorage');
+        $this->shouldHaveType('Sylius\Behat\Service\SharedStorage');
     }
 
     function it_implements_shared_storage_interface()
