+++ src/Sylius/Bundle/CoreBundle/Resources/config/test_services.xml
@@ -16,7 +16,6 @@
            xsi:schemaLocation="http://symfony.com/schema/dic/services
                                http://symfony.com/schema/dic/services/services-1.0.xsd">
     <parameters>
-        <parameter key="sylius.behat.shared_storage.class">Sylius\Component\Core\Test\Services\SharedStorage</parameter>
         <parameter key="sylius.test.factory.user.class">Sylius\Bundle\CoreBundle\Test\Factory\TestUserFactory</parameter>
         <parameter key="sylius.payment_method_name_to_gateway_converter.class">Sylius\Bundle\CoreBundle\Test\Services\PaymentMethodNameToGatewayConverter</parameter>
 
@@ -28,8 +27,6 @@
     </parameters>
 
     <services>
-        <service id="sylius.behat.shared_storage" class="%sylius.behat.shared_storage.class%" />
-
         <service id="sylius.test.factory.user" class="%sylius.test.factory.user.class%">
             <argument type="service" id="sylius.factory.customer" />
             <argument type="service" id="sylius.factory.user" />
